move odf-unrelated functions out of processOdfFile header
