+++ src/tudatpy/estimation/observations_setup/observations_simulation_settings/expose_observations_simulation_settings_factory.cpp
@@ -21,7 +21,7 @@
 #include "scalarTypes.h"
 #include "tudat/simulation/estimation_setup/createObservationModelFactory.h"
 #include "tudat/simulation/estimation_setup/simulateObservations.h"
-#include "tudat/simulation/estimation_setup/processOdfFile.h"
+#include "tudat/simulation/estimation_setup/observationSimulationSettings.h"
 
 namespace tss = tudat::simulation_setup;
 namespace tom = tudat::observation_models;
